Vector DB capabilities
Use QDrant locally (and default on right now - install docker and qdrant - see instructions)

Supports Azure Search and Pinecone (because Semantic Kernel does).

Added ability to upload a pdf or word document to the Vector Store.  It can segment out the document for better search... really just a direct copy of the vector-store-connectors/how-to/vector-store-data-ingestion and added pdf support .